chore: reset unnecessary changes
