+++ packages/dashmate/src/listr/prompts/createPlatformNodeKeyInput.js
@@ -7,11 +7,6 @@ const validateTenderdashNodeKey = require('./validators/validateTenderdashNodeKe
  * @returns {Object}
  */
 function createPlatformNodeKeyInput(options = {}) {
-  let { initial } = options;
-  if (initial === null || initial === undefined) {
-    initial = generateTenderdashNodeKey();
-  }
-
   return {
     type: 'input',
     name: 'platformNodeKey',
@@ -23,7 +18,7 @@ function createPlatformNodeKeyInput(options = {}) {
   automatically generated for you.\n`,
     message: 'Enter Ed25519 node key',
     hint: 'Base64 encoded',
-    initial,
+    initial: options.initial === '' || options.initial ? options.initial : generateTenderdashNodeKey(),
     validate: validateTenderdashNodeKey,
   };
 }
